backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
pydantic-core==2.46.3
requests==2.32.5
deepgram-sdk==3.4.0
python-dotenv==1.2.2
python-multipart==0.0.27
typing-extensions==4.15.0
aiofiles==25.1.0
firebase-admin==7.4.0
google-cloud-storage==3.10.1
httpx==0.28.1
colorama==0.4.6
click==8.3.3
groq
